+++ playbook/app/pb_kits/playbook/pb_draggable/docs/_draggable_default.jsx
@@ -1,91 +1,37 @@
 import React, { useState } from "react";
 import {
-  Flex,
-  Background,
-  Caption,
-  Card,
-  FlexItem,
-  Badge,
-  Avatar,
-  Title,
-  Body,
+ SelectableList,
   Draggable,
   DraggableProvider,
 } from "../../";
-
-// Initial groups to drag between
-const containers = ["To Do", "In Progress", "Done"];
 
 // Initial items to be dragged
 const data = [
   {
     id: "1",
-    container: "To Do",
-    title: "Task 1",
-    description: "Bug fixes",
-    assignee_name: "Terry Miles",
-    assignee_img: "https://randomuser.me/api/portraits/men/44.jpg",
+    container: "List",
+    text: "Task 1",
   },
   {
     id: "2",
-    container: "To Do",
-    title: "Task 2",
-    description: "Documentation",
-    assignee_name: "Sophia Miles",
-    assignee_img: "https://randomuser.me/api/portraits/women/8.jpg",
+    container: "List",
+    text: "Task 2",
   },
   {
     id: "3",
-    container: "In Progress",
-    title: "Task 3",
-    description: "Add a variant",
-    assignee_name: "Alice Jones",
-    assignee_img: "https://randomuser.me/api/portraits/women/10.jpg",
+    container: "List",
+    text: "Task 3",
   },
   {
     id: "4",
-    container: "To Do",
-    title: "Task 4",
-    description: "Add jest tests",
-    assignee_name: "Mike James",
-    assignee_img: "https://randomuser.me/api/portraits/men/8.jpg",
-  },
-  {
-    id: "5",
-    container: "Done",
-    title: "Task 5",
-    description: "Alpha testing",
-    assignee_name: "James Guy",
-    assignee_img: "https://randomuser.me/api/portraits/men/18.jpg",
-  },
-  {
-    id: "6",
-    container: "In Progress",
-    title: "Task 6",
-    description: "Release",
-    assignee_name: "Sally Jones",
-    assignee_img: "https://randomuser.me/api/portraits/women/28.jpg",
+    container: "List",
+    text: "Task 4",
   },
 ];
 
 const DraggableDefault = (props) => {
   const [initialItems, setInitialItems] = useState(data);
 
-  const badgeText = (container) => {
-    return container === "To Do"
-      ? "queue"
-      : container === "In Progress"
-      ? "progress"
-      : "done";
-  };
-
-  const badgeColor = (container) => {
-    return container === "To Do"
-      ? "warning"
-      : container === "In Progress"
-      ? "primary"
-      : "success";
-  };
 
   return (
     <DraggableProvider>
@@ -94,77 +40,32 @@ const DraggableDefault = (props) => {
           setInitialItems(items)} 
           {...props}
       >
-        <Flex
-            justifyContent="center"
-            margin="xs"
-            padding="xs"
-        >
-          {containers?.map((container) => (
             <Draggable.Container 
-                container={container}
-                key={container} 
+                container="List"
             >
-              <Background
-                  backgroundColor="white"
-                  padding="sm"
-              >
-                <Caption textAlign="center">{container}</Caption>
-                <Flex 
-                    alignItems="stretch"
-                    orientation="column" 
-                >
+                    <SelectableList variant="checkbox">
                   {initialItems
-                    .filter((item) => item.container === container)
                     .map(
                       ({
-                        assignee_img,
-                        assignee_name,
-                        description,
                         id,
-                        title,
+                        text,
                       }) => (
                         <Draggable.Item 
-                            container={container}
+                            container="List"
                             id={id} 
                             key={id}
                         >
-                          <Card
-                              marginBottom="sm"
-                              padding="sm"
-                          >
-                            <Flex justify="between">
-                              <FlexItem>
-                                <Flex>
-                                  <Avatar
-                                      imageUrl={assignee_img}
-                                      name={assignee_name}
-                                      size="xxs"
-                                  />
-                                  <Title paddingLeft="xs" 
-                                      size={4}
-                                  >
-                                    {title}
-                                  </Title>
-                                </Flex>
-                              </FlexItem>
-                              <Badge
-                                  rounded
-                                  text={badgeText(container)}
-                                  variant={badgeColor(container)}
-                              />
-                            </Flex>
-                            <Body paddingTop="xs"
-                                text={description} 
-                            />
-                          </Card>
+                        <SelectableList.Item
+                            label={text}
+                            name={id}
+                            value={id}
+                        />
+                          
                         </Draggable.Item>
                       )
                     )}
-                </Flex>
-              </Background>
+                    </SelectableList>
             </Draggable.Container>
-          ))}
-        </Flex>
       </Draggable>
     </DraggableProvider>
   );
