+++ playbook/app/pb_kits/playbook/pb_draggable/docs/index.js
@@ -1,1 +1,2 @@
 export { default as DraggableDefault } from './_draggable_default.jsx'
+export { default as DraggableMultipleContainers } from './_draggable_multiple_containers.jsx'
